+++ code/src/components/TodoList.js
@@ -1,27 +1,14 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
+import { useSelector } from 'react-redux'
 
-import todos from '../reducers/todos'
+import TodoItem from './TodoItem'
 
 const TodoList = () => {
   const items = useSelector((store) => store.todos.items)
-
-  const dispatch = useDispatch()
-
   return (
     <div className="todo-list">
-      {items.map((todo) => (  // mutability -> (todo, index)
-        <div key={todo.id} className="todo-item">
-          <p>{todo.description}</p>
-          <input
-            type="checkbox"
-            checked={todo.isComplete}
-            onChange={() => dispatch(todos.actions.toggleComplete(todo.id))}
-          /> 
-          <button onClick={() => dispatch(todos.actions.removeTodo(todo.id))} className="delete-button">
-            delete
-          </button> 
-        </div>
+      {items.map((todo) => (
+        <TodoItem {...todo} key={todo.id} />
       ))}
     </div>
   )
