+++ code/src/index.css
@@ -31,7 +31,7 @@ code {
 }
 
 .todo-list {
-  border: 2px solid #276fbf;
+  border: 2px solid #183059;
   text-align: center;
   margin: 30px;
   background-color: white;
